+++ lib/haml_lint/linter/leading_comment_space.rb
@@ -6,7 +6,7 @@ module HamlLint
     def visit_haml_comment(node)
       # Skip if the node spans multiple lines starting on the second line,
       # or starts with a space
-      return if node.text.match(/\A(\s*|\s\S.*)$/)
+      return if node.text.match(/\A(\s*|\s+\S.*)$/)
 
       add_lint(node, 'Comment should have a space after the `#`')
     end